Greatly simplify base
Base is narrower (same with as PS2) with just a ridge on the back and a "nub"
on the front. The sid ridges are gone, as the legs will prevent sideways
movement. There is also a void in the back (to save material & printing time)
and a floor on the bottom of the vent area (to add strength).